Update README.md
Change: Remove the 'read-at-version' requirement.

Motivation: Currently, the 'read-at-version' requirement offers minimal value. The Outdated status does not return the old value, and NotAvailableYet is not very useful. Therefore, it is recommended to either remove this requirement, as this PR proposes, or upgrade it to return the previously stored value.